refactoring to custom hook
